Ignore missing 450um source catalogs
For some processing jobs, CUPID can not produce a catalog from the
450um data.  So for now ignore this error.  In future we will have
to handle this more comprehensively, e.g. by knowing which jobs we
expect to fail in this way, by recording a warning of some kind,
or by making sure to delete from VO space any old catalogs which are
no longer generated.