Use obs. info for submitting calibration jobs
Using the output file names to determine the source was fine for
submitting calibration jobs once all of the individual reductions
were complete, but does not work when there are newly-added reduction
jobs in the database.  Therefore use the observation info instead.
(Initially output files were preferred because all of the job
information could be retrieved with one call to "find_jobs".  Now
we also need to call "get_obs_info" for each job.)

Additionally we need to convert source names into the "safe" identifier
strings used by the Transient Survey.  To do this, open another
Taco connection, to Python 3 (instead of Python 2 as is used for jsa_proc)
in order to be able to import the jsa_proc_transient.dr.safe_object_name
function.