Build list of "trigger" values by position
Since we iterate over the fluxes in reverse order, we can't
just append to the triggers list.  Now prepare a list full of
zero value placeholders and then assign to it by positions. This
is so that the order in which the fluxes are processed does not
matter.  (The alternative would have been to reverse the triggers
list again before adding it to the table row, but that would
assume processing in reverse order.)